k8s-container/dockerdir/requirements.txt
azure-keyvault-secrets==4.7.0
msal==1.22.0
psycopg2-binary==2.9.7